MultiSelect dropdown and Anatomic model improvements
Created a static list of muscles and added a new migration for this.

Added SyncFusion library in order to use the MultiSelect dropdown control in the WorkoutOverviewPage.

Saved the associated muscle groups for every exercise after selection is finished in the dropdown.

Added a default image for each muscle group in the AnatomicModel page.

When the user clicks on an image it is displayed the exercises that can be done in order to stimulate the selected muscle.

Then the user can add the from that page any exercise in any of his workout splits and then in it's corresponding workouts