Added anatomic model
Added photos for each muscle group in the anatomic model page.

Added a toggle button to change the perspective of the anatomic model.

The model is interactive so the user can click on almost every muscle and get the associated exercises.

Added the username field in the manage account page.

Installed sync fusion charts for the history page