Landing page and UI improv
Changed color of the navbar's gradient.

Created an index page which autoredirects to home page if the user is not logged in, else the user is redirected to manage account.

Added home item in the navbar menu.

Removed any padding for every new page in order to fix background for landing page.

Added indentAccount class in app.css for register and login page.